Add training/eval pipeline, dataset, and LoRA
Introduce a full training/evaluation pipeline and dataset support for the DACL10K .npy format. Key changes: add Faz 4 training report (docs); tweak requirements formatting. Implement DACL10KDataset to read .npy images/masks, produce SAM3-processor-ready inputs (pixel_values, tokenized prompt, resized 288x288 ground-truth mask) and support sample limiting. Implement training and evaluation logic: mixed-precision forward/backward, BCEWithLogitsLoss, IoU computation, logging, DataLoader usage, saving/loading LoRA adapters. Update config (batch size, mask output size, text prompt, max samples). Load base SAM3 model in float16 to reduce VRAM. Update run_train/run_eval scripts to wire dataset, DataLoader, PEFT adapter loading, and to run train/eval end-to-end.

+++ src/model.py
@@ -37,6 +37,7 @@ Bu dosyada iki fonksiyon var:
 - load_processor() → görselleri modele uygun hale getiren processor'ı yükler
 """
 
+import torch
 from transformers import Sam3Model, Sam3Processor
 
 from src.config import Config
@@ -56,7 +57,9 @@ def load_model():
     log(f"Model yükleniyor: {Config.MODEL_NAME}")
 
     # Modeli HuggingFace'ten indir ve yükle
-    model = Sam3Model.from_pretrained(Config.MODEL_NAME)
+    # torch_dtype=float16 → model yarı hassasiyette yüklenir
+    # Bu sayede: yükleme ~2x hızlı, VRAM kullanımı ~yarıya düşer
+    model = Sam3Model.from_pretrained(Config.MODEL_NAME, torch_dtype=torch.float16)
 
     # Modeli doğru cihaza taşı (GPU varsa GPU, yoksa CPU)
     model = model.to(Config.DEVICE)
